Implement averaging and octree scheme configuration in api

+++ OctreeApi/OctreeLibrary/include/types.cuh
@@ -9,7 +9,7 @@
 
 struct Averaging
 {
-    float r, g, b;
+    uint32_t r, g, b;
     uint32_t pointCount;
 };
